+++ .script/buildManifest.js
@@ -2,11 +2,15 @@ console.log("start generate manifest")
 const fs = require('fs');
 
 const buildEntryFromEnv = process.env.PACKAGE_DIR;
+console.log(process.env.VERSION)
+const VERSION = process.env.VERSION
 
 const packageConfig = require(buildEntryFromEnv ? `${buildEntryFromEnv}/package.json` : 'codingIdePlugin/package.json');
 const mapPackage = require('./mapPackage');
 
-const version = packageConfig.codingIdePackage.version || packageConfig.version
+
+packageConfig.codingIdePackage.version = VERSION || packageConfig.codingIdePackage.version || packageConfig.version
+const version = packageConfig.codingIdePackage.version
 const newPackage = mapPackage(packageConfig)
 fs.writeFile(`dist/${version}/manifest.json`, JSON.stringify(newPackage, null, 4), 
 function(err) {